+++ src/euroeval/benchmark_modules/hf.py
@@ -908,7 +908,9 @@ def _load_model_from_pretrained(
                 )
                 model_kwargs["ignore_mismatched_sizes"] = True
                 continue
-            raise InvalidModel(str(e)) from e
+            raise InvalidModel(
+                f"The model {model_id!r} could not be loaded. The error was {e!r}."
+            ) from e
         except (TimeoutError, RequestError):
             log(
                 f"Couldn't load the model {model_id!r}. Retrying.",
@@ -1340,7 +1342,7 @@ def setup_model_for_question_answering(model: "PreTrainedModel") -> "PreTrainedM
             token_type_embeddings.weight.data = torch.cat(
                 tensors=(
                     token_type_embedding_tensor,
-                    torch.rand_like(tensor=token_type_embedding_tensor),
+                    torch.rand_like(token_type_embedding_tensor),
                 ),
                 dim=0,
             )