+++ src/euroeval/dataset_configs/german.py
@@ -117,6 +117,7 @@ MULTILOKO_DE_CONFIG = DatasetConfig(
     val_split=None,
 )
 
+
 # Unofficial datasets ###
 
 MMLU_DE_CONFIG = DatasetConfig(